Ported LogMeta and refactored CheckpointMeta out + moved RawCommittedMessage
