+++ rust/abacus-core/src/types/mod.rs
@@ -1,14 +1,16 @@
 use ethers::types::{H256, U256};
 
 mod checkpoint;
-mod messages;
+mod log_metadata;
+mod message;
 
 /// Unified 32-byte identifier with convenience tooling for handling
 /// 20-byte ids (e.g ethereum addresses)
 pub mod identifiers;
 
 pub use checkpoint::*;
-pub use messages::*;
+pub use log_metadata::*;
+pub use message::*;
 
 use crate::{AbacusError, Decode, Encode};
